Fixed issue with modification handler on inventories not picking up the correct number of items.

+++ src/main/java/theking530/staticpower/tileentities/components/items/UpgradeInventoryComponent.java
@@ -94,7 +94,7 @@ public class UpgradeInventoryComponent extends InventoryComponent {
 		}
 
 		public float getUpgradeWeight() {
-			return (float) stack.getCount() / stack.getMaxStackSize();
+			return (float) stack.getCount() / (float) stack.getMaxStackSize();
 		}
 
 	}
